+++ image.cpp
@@ -86,7 +86,7 @@ bool Image::readImage(const char * filename) {
 	this->_height = height;
 	std::vector<unsigned char> imageC(width * height * 3);
 	int t = fread(&(imageC[0]), width * height * 3, 1, file);
-	for (int i = 0; i < imageC.size(); i++) {
+	for (size_t i = 0; i < imageC.size(); i++) {
 		_image.push_back((float)imageC[i] / 255.0f);
 	}
 	printf("Loaded texture %s\n", filename);
